reverted back to background image format
The plan is to create a temporary image clone when the crop button is hit. The user will be able to crop the image in this clone. once the cropping is finished, the dimensions of the crop will be used to create a css sprite using the background image already in existence.