Images are now able to be cropped
Unfortunately they cannot be enlarged after they are cropped and they do not retain the previous dimensions of the resized div if it happened to be resized beforehand. This is due to the way CSS sprites function. The effect that it has is basically hiding the entire background image, and then creating a cutout within that hidden background image that shows only part of it. If we try to resize the div, the cutout gets larger, thus showing more of the background image - the background image itself does not resize. I am trying to come up with a way to enlarge the background image size within css at a proprortion of how much of the image has been cropped out.