resize functionality now disabled on cropped images
This is due to the way that the cropping works. Resizing only resizes the cutout of the background image, not the background image itself. so currently, resizing a cropped image only ends up resizing the "Crop." In other words, by resizing, you are undoing the crop effect and restoring the size of the image. There may be a fix down the road, but for now, resizing is disabled.