reset to native on images & createimg btn change
Create image no longer has the user input a url. Instead a placeholder image is created. When clicked, this placeholder image summons the edit window where a url can be pasted in. This will automatically change the image src and bring in the new image at its native resolution.

The images can be reset to their native size as well.

I edited the sources.md to reflect the new addition of the placeholder image