Refresh tree now runs every 15 seconds
part of the slow browser experience could have been coming from the way that the tree refresh function was being called. Every time the data_update function was run, the tree refresh function was being called which is a recurssive function that rebuilds the tree using the data object. It is very possible that this was the cause of the slow down since sometimes it was being called up to 3 times a second. I have since changed the refreshtree function to be called every 15 seconds to help the problem. I have run the same instance of wsWebdisplay for over an hour and firefox has not used more than 25% of the cpu and averages around 8-12%.