wrote a function that populates unit change...
selection list. the function searches within the object array that tracks cells for an id (which is based as an argument). With this ID it is able to find type, units, typeUnits. based on type, it will populate the dropdown selection list with values from a corresponding array. The next step will be a function that converts the value of the cell based on the selection from the user.