Still working on implementing cam
As long as I am doing this, I am going to try to allow more than one of each cam on the page. Currently this is not possible due to the way this was set up in the very beginning with one cam. Currently, the application brings in every cam it can find initially and loads them into an array. I believe this would prevent new cams from coming in without a reload of the page.