log is functioning. cannot add dynamically via jstree.
Currently the way that it works is the object contains a doubly linked list. when it reaches the number set by the logLimit property, it starts removing the head of the list and deleting elements off of the front. The number on the left-hand side is a timestamp rounded to the nearest second, and the number on the right is the data.