wrote function that updates portions of data dynamically
Currently the tree is still updating inefficiently. The way jstree works is it builds the tree from json and requires certain values to be set for each node. These values are the text, and parent of each node. This is why it is processing the entire data object every time in iterate stations. What I might be able to do is change it so that individual data can be changed within the jstree.