last data received timer now takes into account
hibernation. it counts from the last timestamp created instead of a timer. This allows it to retain the time that it went to sleep and count from that time when it wakes up. I also made it so tooltips are disabled when dragging - have yet to do the same for resizing